+++ .gitignore
@@ -110,3 +110,5 @@ broadcast
 
 # interfaces
 contracts/interfaces/
+
+TODO.*
