--- codechecks.yml
@@ -1,2 +1,0 @@
-checks:
-  - name: eth-gas-reporter/codechecks
