+++ config/networks.ts
@@ -25,7 +25,7 @@ const hardhat: NetworksUserConfig['hardhat'] = {
   gasPrice: 3,
   chainId: 9001,
   accounts,
-  loggingEnabled: Boolean(LOG_HARDHAT_NETWORK),
+  loggingEnabled: LOG_HARDHAT_NETWORK,
   allowUnlimitedContractSize: true,
   tags: ['test'],
   saveDeployments: true,
@@ -40,7 +40,7 @@ const localhost: NetworkUserConfig = {
     accounts: { mnemonic: MNEMONIC },
   }),
   allowUnlimitedContractSize: true,
-  loggingEnabled: Boolean(LOG_HARDHAT_NETWORK),
+  loggingEnabled: LOG_HARDHAT_NETWORK,
   url: 'http://127.0.0.1:8545',
   tags: ['test'],
   saveDeployments: true,
