+++ package.json
@@ -16,7 +16,6 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@codechecks/client": "^0.1.12",
     "@ethersproject/abstract-signer": "5.7.0",
     "@ethersproject/properties": "5.7.0",
     "@ethersproject/providers": "5.7.2",
@@ -29,18 +28,19 @@
     "@openzeppelin/contracts": "4.7.3",
     "@openzeppelin/contracts-upgradeable": "4.7.3",
     "@openzeppelin/hardhat-defender": "^1.8.1",
-    "@openzeppelin/hardhat-upgrades": "1.22.0",
+    "@openzeppelin/hardhat-upgrades": "1.22.1",
     "@typechain/ethers-v5": "^10.2.0",
     "@typechain/hardhat": "^6.1.5",
-    "@types/chai": "^4.3.3",
+    "@types/chai": "^4.3.4",
     "@types/chai-as-promised": "^7.1.5",
     "@types/cli-progress": "^3.11.0",
+    "@types/debug": "^4.1.7",
     "@types/dotenv-defaults": "^2.0.1",
-    "@types/proper-lockfile": "^4.1.1",
-    "@types/debug": "^4.1.5",
     "@types/dotenv-parse-variables": "^2.0.1",
     "@types/fs-extra": "^9.0.13",
     "@types/json-diff": "^0.9.0",
+    "@types/mocha": "^10.0.1",
+    "@types/proper-lockfile": "^4.1.2",
     "chai": "^4.3.6",
     "chai-as-promised": "^7.1.1",
     "chalk": "^4.0",
@@ -55,8 +55,8 @@
     "ethereum-waffle": "^3.4.4",
     "ethernal": "^2.0.2",
     "ethers": "^5.7.2",
-    "fireblocks-defi-sdk": "1.2.5",
-    "fireblocks-sdk": "^2.3.2",
+    "fireblocks-defi-sdk": "1.2.15",
+    "fireblocks-sdk": "^3.1.2",
     "fs-extra": "^10.1.0",
     "handlebars": "^4.7.7",
     "hardhat": "2.12.5",
