--- utils/networks.ts
@@ -1,8 +1,0 @@
-import type { SupportedNetworks } from '@/config/networks';
-import { networks } from '@/config/networks';
-
-export const isSupportedNetwork = (
-  name: SupportedNetworks
-): name is SupportedNetworks => {
-  return Object.keys(networks).includes(name);
-};
